Added additonals classes and extensions
Added CrestronConsoleEx class, which has all the methods and properties of the regular CrestronConsole class, but adds ConsoleCommandResponseLine methods, allows PrintLine to be called with no argument to print an empty line, and redirects Print/PrintLine to ConsoleCommandResponse if on the command thread.

Adds the CrestronIPEnpoint class which adds the AdapterType property.

Adds additional methods to the IPAddressEx and IPAddres extensions.